requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
gunicorn==20.1.0
python-multipart==0.0.5
azure-storage-blob==12.9.0
Pillow==8.3.2